Set global and toolbar UWP theme
Set global theme to "light" to improve appearance of toolbar on physical
devices (this overwrites the user setting of theme, which may make the
app look bad).

Set toolbar theme to "dark" for more appropriate touch feedback/hover
colors of the drawer button matching the dark solid background color of
the navigation bar.