Improve naming of data type config setting
Data type defined in the config file is not necessarily coupled to a
service, as data may be bundled in the app.

The setting describes:
  * the way the data is structured.
  * the way remote services behave for a specific data type. Consumable
service is mostly coupled to a specific data format, e.g. we are bound
to drupal API for "cod" and have to support non-standard caching while for
"googleIO" we would have standard caching and resolval of data files
through a manifest.

In this sense the more abstract term "DATA_TYPE" describes more precisely
the meaning of the setting.