Simplify service configuration
Explicitly specify service endpoints instead of hard-coding them in the
remote module. Those may change when a breaking routing change has been
introduced.

Use the constant naming convention for the map properties as in the rest
of the configuration, as we can make real ES6 constants out of them in
the future.

Remove urijs dependency as we don't need it anymore.

Change-Id: I2705eb21782a8e15e7806c228eb91a4fba13d6a2