Add Cordova script to update versions codes
Before, one had to manually update them before uploading a new build
with the same version to the app store. This was time-consuming and
error prone.

Use ${timestamp} (1) on Android and ${version}+${timestamp} (2,3) on iOS,
where ${timestamp} is an Unix timestamp in minutes and ${version} is the
Cordova widget version.

The used timestamp format will not overflow on Android until 02.18.2160.

On iOS the recommendation for CFBundleVersion values (3) has been
adapted to semantic versioning.

[1]: https://developer.android.com/studio/publish/versioning.html
[2]: http://semver.org, see 10.
[3]: https://developer.apple.com/library/content/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html

Change-Id: I849efe725407c389202141aecb04a65f150a3c0a