Preselect single sessions in the schedule
Reduce cognitive load of the user and make information easier to access
by preselecting sessions, to which no other sessions run in parallel.
Those are mostly keynotes.

This follows the same pattern as fixed blocks which don't have
alternative events at the same time. In comparison to them, however,
single sessions (as other sessions too) can be removed from the schedule
by the user if he choses to do it.

Change-Id: I65f93ebd07173a6a68ba94cb0b9d03d7c0607de3