Simplify drawer on Android
Make login button regular item of the drawer list. Remove drawer account
mode, which was triggered by tapping on the drawer when the user was
logged in. It revealed a "Logout" item.

Previously, login item was highlighted and placed on top of the drawer.
This might have been confusing, as profile doesn't have that high of a
priority in this app. Highlighting it may even had adverse effect,
causing it to be missed, as it didn't look like a drawer list item
(see #31).

Furthermore, the drawer account mode might have been confusing.
Revealing a single "Logout" option in this mode was an overkill. Instead
make it always visible in the drawer if the user was logged in.

Fix #31.

Change-Id: I05d21f1656203f336b41cf83790c89eb5b60ea4c