Improve voting UX
Now no "Please login to vote" label will be shown, but a "GIVE FEEDBACK"
button, which opens the login page if the user was not logged in.

This change also includes minor cleanups and change of handling of login
events, which broke encapsulation. Now instead of triggering events on
various widgets, the login service will emit events, which can be handled
by all interested widgets.

Introduce an error logger, which should be used for all promises, where
no errors are caught.

Change-Id: Id875c21f2b10845a6e2e432d55b00189cb3b5bf2