Improve data refresh performance
Simplify comparison of timestamps when getting sessions in a timeframe.
Previously when comparing timestamps, moment objects were constructed
under hood. TimezonedDate is used to pin a timestamp to a timezone
when presenting it. This is not relevant for comparison of dates, and
was resource intensive.

Make sure ConferenceDataProvider.get() only requests one result at once.
Previously, a few parts of the UI would have called get() at the same
time, resulting in the result being requested and processed several
times, costing unneeded power and bandwidth.