Include basic documentation in README
Document running the app and include basic information about
customization.

Include app screenshots to make the README more descriptive. In the
Android screenshot, the drawer revealing the logo placeholder is
displayed with the "Tracks" entry selected. On iOS, the Tracks page is
shown.

Screenshots currently display EclipseCon NA 2015 data. In the future we
should develop a test server with dummy data instead of using concrete
conference data in the framework project.

Change-Id: Ib5d073832cb89a9dc8bd68f4d4453c727a33f61b