Simplify and improve map page
jQuery and the panzoom plugin were not used at all due to a syntax error
in the initialization script. Remove them and change image layout to
ensure map images will be zoomed out to fully cover the view. It is
better to initially show a full overview of the map instead of a quarter
of it.

Remove constant MAP_BACKGROUND_COLOR which is not used anymore. White
background is a good default for most maps.

Change-Id: I470ab83b92d8dc5b3da9051de9bbdf52025454c4