Simplify navigation
Use TabFolder navigation on all platforms for simplicity and to improve
interaction with top-level pages.

Use the property selectedImage on tabs, which is supported as of 1.9.0
and does not have the timing issues faced when setting the image
manually.

Remove drawer related components and the Navigatable class. Navigatable
was the wrong abstraction for item handling anyway. Trigger
initializeItems() explicitly on tracks and schedule tabs, since they are
the only top-level screens showing items.

Fix #36

Change-Id: Ic4665d5f0be8c26dda9f244ad3012ebd5f69a177