Introduce a Webpack-based TypeScript build configuration
As part of the migration to TS (#39), change the build configuration to
use TypeScript with Webpack (#40). Introduce a TSX-compatible TypeScript
compiler configuration.

Replace JSHint and JSCS by TSLint, which checks both code quality and
style. Its configuration only applies to new TS files.

Bump version to 2.0.0 to accomodate for breaking changes in the app
configuration. Bundled conference data cannot be loaded dynamically by
BundledConferenceData, since expressions cannot be used as dependencies
anymore due to limitation in Webpack. Statically import them in the
configuration instead.

Remove tabris-js-node dependency, which is not used anymore.

Migrating test build configuration to TS to be done in a separate change.

Fix #40