Remove LoadingIndicator
It was not needed. The only extra behavior it provided additional to the
one of Tabris.js' ActivityIndicator was showing an overlay shade
underneath it. This was actually even not necessary at the places of its
usage.

Minor refactoring. Remove appear workaround on map page for Android,
since it is not needed anymore with modern Tabris.js.

Change-Id: I5a64940b5083e1dffa88d115f742324e697f45d5