Remove Input component
It was only used in the LoginPage and it just had a single style
attribute set. Use borderColor instead of background, since the original
intention was to change the borderColor on Android. The background color
should not be changed on Windows at all, so setting background color per
default was a bug.

Change-Id: I34754301bc480db9a2a75e4c56569f87fa225c4a