Simplify schedule item focus
Previously, in order to highlight an item in the schedule, the UI got
iterated in order to find the right CollectionView. This was a very
over-complicated way to do that.

Introduce a lightweight data-binding between items and cells in order to
make this easier.

Items now have a reference to their CollectionView. It is initialized on
CollectionView creation. It is used to select the item tab before
highlight. E.g. associating the item with the cell and calling
cell.parent() would not work in this case, since the item might not have
been associated with a cell yet when the schedule tab gets opened.

The CollectionView reference is also used to access the item cell. It is
used to call highlight() on the cell and await its return value in order
to prevent an animation conflicting with a cell update.

Associate highlight() method call with the session action tap.
Propagating this through the attended session service was overly
complicated, since the focus had to be deferred until the page got
opened, which happens directly on action tap.

Name the method highlight() instead of focus(), since focus() may imply
persistent focus, although the item is only highlighted for a second,
after which the highlight disappears.

Change-Id: Ie4033b290a7f30dd5843b36eac974aff40197793