Improve voting login UX
Previously, a "Login" action was always visible. For many users, it was
unclear what the login is used for. Since login is only used in the
context of giving feedback, only login when tapping on the GIVE FEEDBACK
button of a session as part of the voting process and remove the login
action.

Changes in behavior:

* No login action will be shown anymore. Show a profile action after the
  first login, which shows a page with the current user data and a button
  to logout.
* Initially show empty feedback indicators on the sessions during the
  feedback time to suggest that giving feedback is now possible. After
  giving feedback for a session (and logging in during the process), the
  indicators will update to represent the votes of the user account.

Fix #35

Change-Id: I53bd6b870c8bbc80c1d75a51b03bcdf65b290971