Simplify map zoom
Previously, in order to make the venue map zoomable, it had to be
presented in a WebView, which caused some implementation and performance
overhead. Simplify venue map by using the newly introduced ImageView
zoom feature.

Remove ActivityIndicator from the Map page, since the image is loaded
from a bundled resource and this should not take a considerable amount
of time.

Change-Id: Ia9cfbec0e3537044847d610b505a1a53410c8f12