Introduce neutral feedback option
Previously it was not clear how to submit a neutral vote. It was only
possible by voting without selecting one of the both thumbs, and
deselecting a thumb was not possible. Introduce an explicit neutral
option.

Use buttons with face instead of thumb icons. Faces are commonly used in
survey interfaces.

Fix layout of feedback page title. Since the title text view was not
horizontally bound, on small screens it got cut-off.

Fix #32

Change-Id: I0721055f8757756fca794fa669b4dff816be8836