Inject services with tabris-decorators
Previously, singleton services were passed as constructor parameters
throughout the app, which was very cumbersome.

For simplicity, pass configuration as a default parameter when
instantiating services. Configuration to be made injectable in a
separate change.

Currently, concrete services are injected complying with the COD
feedback API. Dependency inversion and refinement of API to be done when
support for a further feedback service is introduced.

Currently, resolve() is used to resolve most dependencies. To be
migrated to constructor injection when the project is fully converted to
TypeScript.